fix(plugin-workflow): preserve workflow variable labels on reopen
